ENH: Log when peroidic monitor message thread stopped
